Updated award table to load record into award form on edit
