+++ erp-client/src/app/app.component.html
@@ -9,7 +9,6 @@
 
     <ul class="menu-links">
       <li><a mat-button href="/awards">Awards</a></li>
-      <li><a mat-button href="/create-award">Create Award</a></li>
       <li><a mat-button href="/">Employees</a></li>
       <li><a mat-button (click)="handleLogout()">Logout</a></li>
     </ul>
