--- erp-client/src/app/components/award/award-form/award-form.component.scss
@@ -1,78 +1,0 @@
-@import '../../../../styles.scss';
-
-form {
-  display: flex;
-  flex-direction: column;
-}
-
-.form-group {
-  display: flex;
-  justify-content: space-between;
-
-  .col-half {
-    width: 45%;
-  }
-}
-
-
-mat-form-field.mat-form-field {
-  width: 100%;
-}
-
-
-.form-actions {
-  margin-top: 30px;
-}
-
-
-button,
-button:hover {
-  transition: background-color 150ms ease;
-}
-
-button {
-  margin-right: 15px;
-}
-
-button.neutral {
-  background-color: mat-color($mat-grey, 200);
-  color: mat-color($mat-gray, 600);
-  
-  &:hover {
-    background-color: mat-color($mat-grey, 300);
-    color: mat-color($mat-gray, 700);
-  }
-}
-
-
-// I know using !important is a bad practice, but it's
-// the only way I could get the timepicker to use whatever
-// styles are defined in our theme.
-
-::ng-deep .timepicker {
-  .timepicker__header {
-    background-color: mat-color($primary, 400) !important;
-  }
-
-  .clock-face__clock-hand {
-    background-color: mat-color($primary, 400) !important;
-
-    &::after {
-      background-color: mat-color($primary, 400) !important;
-    }
-
-    &::before {
-      border-color: mat-color($primary, 400) !important;
-    }
-  }
-
-  .clock-face__number {
-    > span.active {
-      background-color: mat-color($primary, 400) !important;
-    }
-  }
-
-  button:not(.timepicker-dial__item) {
-    color: mat-color($primary, 400) !important;
-  }
-}
