--- erp-client/src/app/components/award/award-table/award-table.component.html
@@ -1,59 +1,0 @@
-<div id="award-datatable">
-  <div id="pre-table-header">
-    <h1 class="mat-h1">My Awards</h1>
-    <p class="mat-subheading-1">View, edit, or delete awards that you've created.</p>
-    <mat-form-field>
-      <input matInput type="text" (keyup)="filter($event.target.value)" placeholder="Type to filter">
-    </mat-form-field>
-  </div>
-
-  <mat-table [dataSource]="dataSource" matSort>
-    <ng-container matColumnDef="id">
-      <mat-header-cell *matHeaderCellDef mat-sort-header class="col-slim">Id</mat-header-cell>
-      <mat-cell *matCellDef="let award" class="col-slim"> {{ award.id }} </mat-cell>
-    </ng-container>
-
-    <ng-container matColumnDef="awardType">
-      <mat-header-cell *matHeaderCellDef mat-sort-header class="col-wide">Award Type</mat-header-cell>
-      <mat-cell *matCellDef="let award" class="col-wide"> {{ award.awardType.name }} </mat-cell>
-    </ng-container>
-
-    <ng-container matColumnDef="employee">
-      <mat-header-cell *matHeaderCellDef mat-sort-header class="col-wide">Employee</mat-header-cell>
-      <mat-cell *matCellDef="let award" class="col-wide"> {{ award.employee.firstName + ' ' + award.employee.lastName }} </mat-cell>
-    </ng-container>
-
-    <ng-container matColumnDef="awardedDate">
-      <mat-header-cell *matHeaderCellDef mat-sort-header>Awarded Date</mat-header-cell>
-      <mat-cell *matCellDef="let award"> {{ award.awardedDate | date:'shortDate' }} </mat-cell>
-    </ng-container>
-
-    <ng-container matColumnDef="awardedTime">
-      <mat-header-cell *matHeaderCellDef mat-sort-header>Awarded Time</mat-header-cell>
-      <mat-cell *matCellDef="let award"> {{ award.awardedTime | time }} </mat-cell>
-    </ng-container>
-
-    <ng-container matColumnDef="edit">
-      <mat-header-cell *matHeaderCellDef class="col-slim ml">Edit</mat-header-cell>
-      <mat-cell *matCellDef="let award" class="col-slim ml">
-        <button mat-icon-button color="accent" (click)="handleEdit(award.id)">
-          <mat-icon class="mat-18">edit</mat-icon>
-        </button>
-      </mat-cell>
-    </ng-container>
-   
-    <ng-container matColumnDef="delete">
-      <mat-header-cell *matHeaderCellDef class="col-slim">Delete</mat-header-cell>
-      <mat-cell *matCellDef="let award" class="col-slim">
-        <button mat-icon-button color="accent" (click)="handleDelete(award.id)">
-          <mat-icon class="mat-18">delete</mat-icon>
-        </button>
-      </mat-cell>
-    </ng-container>
-
-    <mat-header-row *matHeaderRowDef="displayedColumns"></mat-header-row>
-    <mat-row *matRowDef="let row; columns: displayedColumns;"></mat-row>
-  </mat-table>
-
-  <mat-paginator [length]="length" [pageSizeOptions]="[10, 25, 100]" showFirstLastButtons></mat-paginator>
-</div>
