--- erp-client/src/app/components/award/award-table/award-table.component.scss
@@ -1,12 +1,0 @@
-.col-slim {
-  max-width: 70px;
-}
-
-.col-wide {
-  min-width: 20vw;
-  max-width: 250px;
-}
-
-.ml {
-  margin-left: 10px;
-}
