+++ erp-client/src/app/components/award/award.component.scss
@@ -1,3 +1,146 @@
+@import '../../../styles.scss';
+
+// 
+// award
+// 
+
 .tab-wrapper {
   margin-top: 40px;
+}
+
+
+// 
+// award form
+// 
+
+
+form {
+  display: flex;
+  flex-direction: column;
+}
+
+.form-group {
+  display: flex;
+  justify-content: space-between;
+
+  .col-half {
+    width: 45%;
+  }
+}
+
+
+mat-form-field.mat-form-field {
+  width: 100%;
+}
+
+
+.form-actions {
+  margin-top: 30px;
+}
+
+
+button,
+button:hover {
+  transition: background-color 150ms ease;
+}
+
+button {
+  margin-right: 15px;
+}
+
+button.neutral {
+  background-color: mat-color($mat-grey, 200);
+  color: mat-color($mat-gray, 600);
+  
+  &:hover {
+    background-color: mat-color($mat-grey, 300);
+    color: mat-color($mat-gray, 700);
+  }
+}
+
+
+
+// I know using !important is a bad practice, but it's
+// the only way I could get the timepicker to use whatever
+// styles are defined in our theme.
+
+::ng-deep .timepicker {
+  .timepicker__header {
+    background-color: mat-color($primary, 400) !important;
+  }
+
+  .clock-face__clock-hand {
+    background-color: mat-color($primary, 400) !important;
+
+    &::after {
+      background-color: mat-color($primary, 400) !important;
+    }
+
+    &::before {
+      border-color: mat-color($primary, 400) !important;
+    }
+  }
+
+  .clock-face__number {
+    > span.active {
+      background-color: mat-color($primary, 400) !important;
+    }
+  }
+
+  button:not(.timepicker-dial__item) {
+    color: mat-color($primary, 400) !important;
+  }
+}
+
+
+
+// 
+// award table
+// 
+
+mat-header-cell:first-of-type,
+mat-cell:first-of-type {
+  padding-left: 12px;
+  max-width: 60px;
+}
+
+mat-header-cell:last-of-type,
+mat-cell:last-of-type {
+  padding-right: 12px;
+}
+
+.col-action {
+  max-width: 60px;
+  margin-left: 18px;
+  margin-right: -18px;
+}
+
+.col-wide {
+  width: 20vw;
+  min-width: 96px;
+  max-width: 264px;
+}
+
+.col-2 {
+  flex: 2;
+}
+
+.col-3 {
+  flex: 3;
+}
+
+.col-4 {
+  flex: 4;
+}
+
+.ml {
+  margin-left: 12px;
+}
+
+
+
+.sm-hidden {
+  @media screen and (max-width: 960px) {
+    display: none;
+  }
 }
