+++ erp-client/src/styles.scss
@@ -29,7 +29,7 @@ body {
 //
 
 .container {
-  width: 80%;
+  width: 85%;
   max-width: 1200px;
   margin: 60px auto;
 }
